new headers for axis/coord
