+++ staging_vespalib/src/vespa/vespalib/metrics/CMakeLists.txt
@@ -1,7 +1,9 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(staging_vespalib_vespalib_metrics OBJECT
     SOURCES
+    axis.cpp
     clock.cpp
+    coordinate.cpp
     counter.cpp
     dummy_metrics_collector.cpp
     gauge.cpp
