+++ staging_vespalib/src/vespa/vespalib/metrics/axis.cpp
@@ -1,0 +1,2 @@
+// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+#include "axis.h"
