+++ staging_vespalib/src/vespa/vespalib/metrics/coordinate.h
@@ -1,0 +1,19 @@
+// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+#pragma once
+
+#include <vespa/vespalib/stllike/string.h>
+
+namespace vespalib {
+namespace metrics {
+
+using CoordinateValue = vespalib::string;
+
+class Coordinate {
+    const size_t _coord_idx;
+public:
+    size_t id() const { return _coord_idx; }
+    Coordinate(size_t id) : _coord_idx(id) {}
+};
+
+} // namespace vespalib::metrics
+} // namespace vespalib
